JSON Configuration Support and More
- JSON Configuration Support

bliss initramfs finally has json configuration support.
What this means is that instead of you having to manually
edit the bliss-initramfs specific python hook files to
make your changes (adding module/firmware preferences, etc),
you can now simply modify the /etc/bliss-initramfs/settings.json
file. This means that whenever you update bliss-initramfs, Gentoo's
CONFIG_PROTECT will automatically kick in and allow you to safely
merge any configuration changes with the ones you have locally.

I've also exposed many other settings that may not have been exposed.
Not all settings/variables are exposed since some of them are for
internal use only. If a need arises that we need to expose more,
we can do so.

- Support for copying everything under a specific firmware directory

Before we would only be able to copy either specific firmware files,
or all of the firmware folders. Now we have the option to specify
what firmware folders we want to copy. The benefit of this is that
it still allows us to have a small initramfs, but without having this
all or nothing approach.

- As part of the of the centralization of settings into the settings.json
  file, the autodetection of paths for where your udev provider is located,
  is no more. For people that are not using systemd, you'll have to tweak
  your settings.json so that you can update the path for "/lib/systemd/systemd-udev"
  in the base -> files and also base -> udevPath variables accordingly.

- If copying firmware is enabled, but no files/directories are specified,
  you will no longer get a warning saying that these variables are empty.

- Other styling and documentation changes/updates are included