Copy/Remove calls are no longer using shutil
Bedrock Linux uses FUSE in order to divide
their stratus. For whatever reason, shutil
is freaking out when attempting to recursively
copy a directory that is a directory yet complains
that it's a directory -_-. Switching the implementation
of all of our copy and remove commands to be standard
UNIX processes (cp, rm) to do our operations.